// src/lib/schemas.ts
export const QuestionSchema = z.object({
    id: z.string().optional(),
    materia: z.string({ required_error: "O campo 'materia' é obrigatório." }).trim().min(1, "O campo 'materia' não pode estar vazio."),
    assunto: z.string({ required_error: "O campo 'assunto' é obrigatório." }).trim().min(1, "O campo 'assunto' não pode estar vazio."),
    enunciado: z.string({ required_error: "O campo 'enunciado' é obrigatório." }).trim().min(1, "O campo 'enunciado' não pode estar vazio."),
    alternativas: z.array(z.string()).min(2, "A questão deve ter pelo menos 2 alternativas."),
    correta: z.coerce.number({ invalid_type_error: "O campo 'correta' deve ser um número." }).int(),
    explicacao: z.string({ required_error: "O campo 'explicacao' é obrigatória." }).trim().min(1, "O campo 'explicacao' não pode estar vazio."),
    nivelDificuldade: z.enum(['Fácil', 'Médio', 'Difícil', 'Hardcore']).optional().default('Médio'),
    banca: z.string().trim().optional().nullable(),
    ano: z.coerce.number().optional().nullable(),
    tags: z.array(z.string()).optional().default([]), // tags centrais
    xpCustom: z.coerce.number().optional().nullable(),
    createdAt: z.any().optional().nullable(),
    updatedAt: z.any().optional().nullable(),
    deletedAt: z.any().optional().nullable(),
});
export const QuestionListSchema = z.array(QuestionSchema);